commit 6: - popravljena kretnja padova - popravljeno da loptica kod velike brzine ne prelazi ekran - stavljena bijela boja na lopticu da lakše mogu vidjeti kako se ponaša na koliziji s padovima

+++ ping_pong/game.h
@@ -66,7 +66,7 @@ class Game
 		int desni_rezultat;
 		int menu_position;
 		int message_position;
-		int starting_ball_speed;
+		float starting_ball_speed;
 
 		bool desno;
 		bool gore;
